projects/smartfunnel/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
openai==1.3.7
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0
python-dotenv==1.0.0
lxml==4.9.3
fake-useragent==1.4.0
webdriver-manager==4.0.1
schedule==1.2.0 